requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.36
Werkzeug==2.0.3
WTForms==3.0.1
